Create README.md
I'm a professional photographer. Now it's time when images can be created not only with a camera. I want to keep up with other users in this direction. This is a completely new approach for me and I have a lot to learn. Your site gives an opportunity to understand complex things and it's great.